requirements.txt
asgiref==3.9.1
Django==5.2.5
djangorestframework==3.16.1
mysqlclient==2.2.7
sqlparse==0.5.3
typing_extensions==4.14.1
requests==2.31.0
python-dotenv==1.0.1
django-autoslug==1.9.9
python-jose==3.3.0
django-cors-headers==4.4.0
drf-yasg==1.21.7